Update: 1. Address PR comments, moved new kotlin files to 'common4j/src/main/com/microsoft/identity/common/java/' to keep it consistent with rest of kotlin file location.
