+++ common4j/src/main/com/microsoft/identity/common/java/broker/IBrokerPerformanceMetricsProvider.kt
@@ -1,0 +1,5 @@
+package com.microsoft.identity.common.java.broker
+
+interface IBrokerPerformanceMetricsProvider {
+    var brokerPerformanceMetrics: BrokerPerformanceMetrics
+}
